requirements.txt
# ============================================
# Python Dependencies for Flight Price Pipeline
# ============================================

# Airflow (included in Docker image, listed for reference)
apache-airflow==2.7.0

# Database connectors
psycopg2-binary==2.9.7
sqlalchemy==1.4.49

# Data processing
pandas==2.0.3

# dbt for transformations
dbt-postgres==1.7.0

# Machine Learning
scikit-learn==1.3.0
joblib==1.3.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Streamlit (installed separately in container)
# streamlit==1.28.0

# Airflow providers
apache-airflow-providers-http==4.5.0
